feat: create data service for the application
This commit created the following features:
- StockDataService (with caching and sqlite3 integration)
- SqliteStockRepository (tools needed for StockDataService)

Now, every queried stock is first compared to local cache, then if it's
not there, program tries to retrieve it from the database.

Only if the two above fail, the data for the given ticker symbol is
queried, then cached and stored in the database.